+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-05 10:33:32
-// Commit: feat(test): add bi-directional event bus handler for test
+// Updated at: 2026-05-05 13:08:42
+// Commit: fix(test): fix memory leak by cancelling test subscription
 
 class testComponentState {
-  final String commitRef = '870';
-  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
-  final DateTime timestamp = DateTime.parse('2026-05-05T10:33:32+05:00');
+  final String commitRef = '880';
+  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-05T13:08:42+05:00');
 }